# openapi.json
{"openapi":"3.1.0","info":{"title":"Finance Dashboard API","description":"## Role-aware Finance Backend\n\nA structured REST API for managing financial records with role-based access control.\n\n### Roles\n| Role | Capabilities |\n|---|---|\n| **viewer** | View records and recent activity |\n| **analyst** | viewer + create records + access dashboard analytics |\n| **admin** | Full access — manage users, update/delete records |\n\n### Authentication\nUse `POST /auth/login` to obtain a JWT token, then click **Authorize** and enter: `Bearer <your_token>`","contact":{"name":"Finance Dashboard Team"},"license":{"name":"MIT"},"version":"1.0.0"},"paths":{"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new user","description":"Register a new user account.\n\n- Default role is **viewer** unless explicitly specified.\n- Email must be unique across the system.\n- Password must be at least 6 characters.","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and receive a JWT access token","description":"Authenticate with email and password.\n\nReturns a **JWT bearer token** to use in the `Authorization` header\nfor all protected endpoints.\n\nError cases:\n- `401` — wrong email or password\n- `403` — account is deactivated","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["User Management"],"summary":"Get your own user profile","description":"Return the profile of the currently authenticated user.","operationId":"get_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/":{"get":{"tags":["User Management"],"summary":"[Admin] List all users","description":"**Admin only.** Return all users registered in the system.","operationId":"list_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response List Users Users  Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/{user_id}":{"get":{"tags":["User Management"],"summary":"[Admin] Get a specific user by ID","description":"**Admin only.** Fetch details of a specific user by their ID.","operationId":"get_user_users__user_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["User Management"],"summary":"[Admin] Delete a user","description":"**Admin only.** Permanently delete a user account.\n\n- You cannot delete your own account.","operationId":"delete_user_users__user_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/role":{"put":{"tags":["User Management"],"summary":"[Admin] Update a user's role","description":"**Admin only.** Assign a new role (viewer / analyst / admin) to a user.","operationId":"update_role_users__user_id__role_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/status":{"put":{"tags":["User Management"],"summary":"[Admin] Activate or deactivate a user","description":"**Admin only.** Toggle a user's active/inactive status.","operationId":"update_status_users__user_id__status_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/":{"post":{"tags":["Financial Records"],"summary":"[Analyst / Admin] Create a new financial record","description":"**Analyst / Admin only.** Create a new income or expense record.\n\n- `amount` must be a positive number.\n- `type` must be `income` or `expense`.\n- `date` is the actual transaction date (not today's date automatically).","operationId":"create_record_records__post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Financial Records"],"summary":"[All Roles] List financial records with optional filters","description":"**All authenticated roles.** Retrieve a paginated, filtered list of financial records.\n\nSupports filtering by:\n- `type` — income or expense\n- `category` — partial text match (e.g., \"rent\" matches \"Monthly Rent\")\n- `from_date` / `to_date` — date range filter\nFilters out softly deleted records automatically.","operationId":"list_records_records__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search term for category or description","title":"Q"},"description":"Search term for category or description"},{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/RecordType"},{"type":"null"}],"description":"Filter by type: income or expense","title":"Type"},"description":"Filter by type: income or expense"},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by category (partial match)","title":"Category"},"description":"Filter by category (partial match)"},{"name":"from_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Filter records on or after this date (YYYY-MM-DD)","title":"From Date"},"description":"Filter records on or after this date (YYYY-MM-DD)"},{"name":"to_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Filter records on or before this date (YYYY-MM-DD)","title":"To Date"},"description":"Filter records on or before this date (YYYY-MM-DD)"},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"description":"Number of records to skip (pagination offset)","default":0,"title":"Skip"},"description":"Number of records to skip (pagination offset)"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"description":"Max records to return (1-200)","default":50,"title":"Limit"},"description":"Max records to return (1-200)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecordResponse"},"title":"Response List Records Records  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/records/{record_id}":{"get":{"tags":["Financial Records"],"summary":"[All Roles] Get a single financial record","description":"**All authenticated roles.** Retrieve the full details of a specific record.","operationId":"get_record_records__record_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Financial Records"],"summary":"[Admin] Update an existing financial record","description":"**Admin only.** Partially update a financial record.\n\nOnly fields included in the request body are updated — omitted fields remain unchanged.","operationId":"update_record_records__record_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Financial Records"],"summary":"[Admin] Delete a financial record","description":"**Admin only.** Permanently delete a financial record.","operationId":"delete_record_records__record_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"record_id","in":"path","required":true,"schema":{"type":"integer","title":"Record Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/summary":{"get":{"tags":["Dashboard"],"summary":"[Analyst / Admin] Get financial summary","description":"**Analyst / Admin only.** Return aggregated financial summary:\n\n- `total_income` — sum of all income records\n- `total_expenses` — sum of all expense records\n- `net_balance` — income minus expenses\n- `total_records` — total count of all records","operationId":"summary_dashboard_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SummaryResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/dashboard/categories":{"get":{"tags":["Dashboard"],"summary":"[Analyst / Admin] Get totals grouped by category","description":"**Analyst / Admin only.** Return each category's total amount and record count,\nbroken down by type (income / expense). Useful for pie charts or category analysis.","operationId":"categories_dashboard_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CategoryTotal"},"type":"array","title":"Response Categories Dashboard Categories Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/dashboard/trends":{"get":{"tags":["Dashboard"],"summary":"[Analyst / Admin] Get monthly income and expense trends","description":"**Analyst / Admin only.** Return monthly aggregated income, expenses, and net balance.\nUse `months` to control how many recent months are included (default: 12, max: 24).","operationId":"trends_dashboard_trends_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"months","in":"query","required":false,"schema":{"type":"integer","maximum":24,"minimum":1,"description":"Number of recent months to return","default":12,"title":"Months"},"description":"Number of recent months to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MonthlyTrend"},"title":"Response Trends Dashboard Trends Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/recent":{"get":{"tags":["Dashboard"],"summary":"[All Roles] Get most recent financial activity","description":"**All authenticated roles.** Return the most recently dated financial records.\nUse `limit` to control how many are returned (default: 10, max: 50).","operationId":"recent_dashboard_recent_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Number of recent records to return","default":10,"title":"Limit"},"description":"Number of recent records to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RecentRecord"},"title":"Response Recent Dashboard Recent Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Health"],"summary":"API root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health check","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CategoryTotal":{"properties":{"category":{"type":"string","title":"Category"},"type":{"type":"string","title":"Type"},"total":{"type":"number","title":"Total"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["category","type","total","count"],"title":"CategoryTotal"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MonthlyTrend":{"properties":{"year":{"type":"integer","title":"Year"},"month":{"type":"integer","title":"Month"},"income":{"type":"number","title":"Income"},"expenses":{"type":"number","title":"Expenses"},"net":{"type":"number","title":"Net"}},"type":"object","required":["year","month","income","expenses","net"],"title":"MonthlyTrend"},"RecentRecord":{"properties":{"id":{"type":"integer","title":"Id"},"amount":{"type":"number","title":"Amount"},"type":{"type":"string","title":"Type"},"category":{"type":"string","title":"Category"},"date":{"type":"string","format":"date","title":"Date"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["id","amount","type","category","date","description"],"title":"RecentRecord"},"RecordCreate":{"properties":{"amount":{"type":"number","exclusiveMinimum":0.0,"title":"Amount","description":"Amount must be a positive number","examples":[2500.0]},"type":{"allOf":[{"$ref":"#/components/schemas/RecordType"}],"examples":["income"]},"category":{"type":"string","maxLength":100,"minLength":1,"title":"Category","examples":["Salary"]},"date":{"type":"string","format":"date","title":"Date","examples":["2025-06-15"]},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description","examples":["Monthly salary deposit"]}},"type":"object","required":["amount","type","category","date"],"title":"RecordCreate"},"RecordResponse":{"properties":{"id":{"type":"integer","title":"Id"},"amount":{"type":"number","title":"Amount"},"type":{"$ref":"#/components/schemas/RecordType"},"category":{"type":"string","title":"Category"},"date":{"type":"string","format":"date","title":"Date"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"created_by":{"type":"integer","title":"Created By"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["id","amount","type","category","date","description","created_by","created_at","updated_at"],"title":"RecordResponse"},"RecordType":{"type":"string","enum":["income","expense"],"title":"RecordType"},"RecordUpdate":{"properties":{"amount":{"anyOf":[{"type":"number","exclusiveMinimum":0.0},{"type":"null"}],"title":"Amount"},"type":{"anyOf":[{"$ref":"#/components/schemas/RecordType"},{"type":"null"}]},"category":{"anyOf":[{"type":"string","maxLength":100,"minLength":1},{"type":"null"}],"title":"Category"},"date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date"},"description":{"anyOf":[{"type":"string","maxLength":500},{"type":"null"}],"title":"Description"}},"type":"object","title":"RecordUpdate"},"RoleUpdate":{"properties":{"role":{"allOf":[{"$ref":"#/components/schemas/UserRole"}],"description":"New role to assign"}},"type":"object","required":["role"],"title":"RoleUpdate"},"StatusUpdate":{"properties":{"is_active":{"type":"boolean","title":"Is Active","description":"Set to true to activate, false to deactivate"}},"type":"object","required":["is_active"],"title":"StatusUpdate"},"SummaryResponse":{"properties":{"total_income":{"type":"number","title":"Total Income"},"total_expenses":{"type":"number","title":"Total Expenses"},"net_balance":{"type":"number","title":"Net Balance"},"total_records":{"type":"integer","title":"Total Records"}},"type":"object","required":["total_income","total_expenses","net_balance","total_records"],"title":"SummaryResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"$ref":"#/components/schemas/UserResponse"}},"type":"object","required":["access_token","user"],"title":"TokenResponse"},"UserCreate":{"properties":{"name":{"type":"string","maxLength":100,"minLength":2,"title":"Name","examples":["Alice Admin"]},"email":{"type":"string","format":"email","title":"Email","examples":["alice@example.com"]},"password":{"type":"string","minLength":6,"title":"Password","examples":["secret123"]},"role":{"anyOf":[{"$ref":"#/components/schemas/UserRole"},{"type":"null"}],"description":"Defaults to 'viewer' if not provided","default":"viewer"}},"type":"object","required":["name","email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email","examples":["admin@finance.com"]},"password":{"type":"string","title":"Password","examples":["admin123"]}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"role":{"$ref":"#/components/schemas/UserRole"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","email","role","is_active","created_at"],"title":"UserResponse"},"UserRole":{"type":"string","enum":["viewer","analyst","admin"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}